fixed memory leak

+++ src/Circle.h
@@ -51,18 +51,17 @@ namespace freeture
 
     private :
 
-
         cv::Point   mPos;     // Center position.
-        int     mRadius;
+        float       mRadius;
 
     public:
 
-        Circle(cv::Point center, int radius) :mPos(center), mRadius(radius) {
+        Circle(cv::Point center, float radius) :mPos(center), mRadius(radius) {
 
         }
 
         cv::Point getCenter() { return mPos; };
-        int getRadius() { return mRadius; };
+        float getRadius() { return mRadius; };
 
         bool computeDiskSurfaceIntersection(Circle c2,
             float& surfaceCircle1,
@@ -83,12 +82,12 @@ namespace freeture
             intersectedSurface = 0.0;
 
 
-            if(enableDebug) circle(map, mPos, mRadius, cv::Scalar(0,255,0));
-            if(enableDebug) circle(map, c2.getCenter(), c2.getRadius(), cv::Scalar(0,0,255));
+            if(enableDebug) circle(map, mPos, (int)mRadius, cv::Scalar(0,255,0));
+            if(enableDebug) circle(map, c2.getCenter(), (int)c2.getRadius(), cv::Scalar(0,0,255));
 
 
             // Distance between two circles centers
-            float distPcNc = sqrt(pow((mPos.x - c2.getCenter().x), 2) + pow((mPos.y - c2.getCenter().y), 2));
+            double distPcNc = sqrt(pow((mPos.x - c2.getCenter().x), 2) + pow((mPos.y - c2.getCenter().y), 2));
 
             // No intersections.
             if (distPcNc > c2.getRadius() + mRadius) {
@@ -129,23 +128,23 @@ namespace freeture
 
                 if (mPos.y != c2.getCenter().y) {
 
-                    float N = (pow(R1, 2) - pow(R0, 2) - pow(x1, 2) + pow(x0, 2) - pow(y1, 2) + pow(y0, 2)) / (2 * (y0 - y1));
-                    float A = pow((x0 - x1) / (y0 - y1), 2) + 1;
-                    float B = 2 * y0 * ((x0 - x1) / (y0 - y1)) - 2 * N * ((x0 - x1) / (y0 - y1)) - 2 * x0;
-                    float C = pow(x0, 2) + pow(y0, 2) + pow(N, 2) - pow(R0, 2) - 2 * y0 * N;
+                    double N = (pow(R1, 2) - pow(R0, 2) - pow(x1, 2) + pow(x0, 2) - pow(y1, 2) + pow(y0, 2)) / (2 * (y0 - y1));
+                    double A = pow((x0 - x1) / (y0 - y1), 2) + 1;
+                    double B = 2 * y0 * ((x0 - x1) / (y0 - y1)) - 2 * N * ((x0 - x1) / (y0 - y1)) - 2 * x0;
+                    double C = pow(x0, 2) + pow(y0, 2) + pow(N, 2) - pow(R0, 2) - 2 * y0 * N;
                     double delta = std::sqrt(pow(B, 2) - 4 * A * C);
 
                     //cout << delta << endl;
 
                     if (delta > 0) {
 
-                        float resX1 = (-B - delta) / (2 * A);
-                        float resX2 = (-B + delta) / (2 * A);
+                        double resX1 = (-B - delta) / (2 * A);
+                        double resX2 = (-B + delta) / (2 * A);
 
-                        float resY1 = N - resX1 * ((x0 - x1) / (y0 - y1));
-                        float resY2 = N - resX2 * ((x0 - x1) / (y0 - y1));
+                        double resY1 = N - resX1 * ((x0 - x1) / (y0 - y1));
+                        double resY2 = N - resX2 * ((x0 - x1) / (y0 - y1));
 
-                        if(enableDebug) line(map, cv::Point(resX1,resY1 ), cv::Point(resX2,resY2 ), cv::Scalar(255,255,255), 1, cv::LINE_AA);
+                        if(enableDebug) line(map, cv::Point((int)resX1, (int)resY1 ), cv::Point((int)resX2, (int)resY2 ), cv::Scalar(255,255,255), 1, cv::LINE_AA);
 
 
                         // Circle1 more inside the other
@@ -225,22 +224,22 @@ namespace freeture
                 }
                 else {
 
-                    float x = (pow(R1, 2) - pow(R0, 2) - pow(x1, 2) + pow(x0, 2)) / (2 * (x0 - x1));
-                    float A = 1.0;
-                    float B = -2 * y1;
-                    float C = pow(x1, 2) + pow(x, 2) - 2 * x1 * x + pow(y1, 2) - pow(R1, 2);
+                    double x = (pow(R1, 2) - pow(R0, 2) - pow(x1, 2) + pow(x0, 2)) / (2 * (x0 - x1));
+                    double A = 1.0;
+                    double B = -2 * y1;
+                    double C = pow(x1, 2) + pow(x, 2) - 2 * x1 * x + pow(y1, 2) - pow(R1, 2);
 
                     double delta = std::sqrt(pow(B, 2) - 4 * A * C);
 
                     if (delta > 0) {
 
-                        float resY1 = (-B - delta) / (2 * A);
-                        float resY2 = (-B + delta) / (2 * A);
+                        double resY1 = (-B - delta) / (2 * A);
+                        double resY2 = (-B + delta) / (2 * A);
 
-                        float resX1 = (pow(R1, 2) - pow(R0, 2) - pow(x1, 2) + pow(x0, 2) - pow(y1, 2) + pow(y0, 2)) / (2 * (x0 - x1));
-                        float resX2 = (pow(R1, 2) - pow(R0, 2) - pow(x1, 2) + pow(x0, 2) - pow(y1, 2) + pow(y0, 2)) / (2 * (x0 - x1));
+                        double resX1 = (pow(R1, 2) - pow(R0, 2) - pow(x1, 2) + pow(x0, 2) - pow(y1, 2) + pow(y0, 2)) / (2 * (x0 - x1));
+                        double resX2 = (pow(R1, 2) - pow(R0, 2) - pow(x1, 2) + pow(x0, 2) - pow(y1, 2) + pow(y0, 2)) / (2 * (x0 - x1));
 
-                        if(enableDebug) line(map, cv::Point(resX1,resY1 ), cv::Point(resX2,resY2 ), cv::Scalar(255,255,255), 1, cv::LINE_AA);
+                        if(enableDebug) line(map, cv::Point((int)resX1,(int)resY1 ), cv::Point((int)resX2, (int)resY2 ), cv::Scalar(255,255,255), 1, cv::LINE_AA);
 
 
                         // Circle neg more inside the other
